requirements.txt
# 基础依赖
aiofiles>=24.1.0
aiopath>=0.6.11
fastapi>=0.110.2
httpx>=0.28.1
loguru>=0.7.3
numpy>=2.2.3
opendal>=0.45.16
pillow>=11.1.0
pydantic>=2.10.6
pydantic-settings>=2.8.1
python-multipart>=0.0.9
pyyaml>=6.0.2
qdrant-client>=1.9.2
rich>=13.9.4
typer>=0.15.2
uvicorn[standard]>=0.34.0
wcmatch>=10.0

# CPU版本依赖
torch>=2.6.0
torchvision
easypaddleocr>=0.3
transformers>=4.49.0

# 开发依赖(可选)
# pylint>=3.3.4
# pytest>=8.3.5
# pytest-asyncio>=0.25.3
# pytest-cov>=6.0.0